requirements-fastapi.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
overpy==0.7
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.23.2
python-dotenv==1.0.0
